Retire the SOC classification tools
As confirmed with the business, these SOC tools are not the current version
of the SOC code lists and removing them will not have a large impact on
users. The banner that was displayed for over a month before this
decommissioning resulted in no concerns raised by users about us
retiring these tools. Based on this these tools are now being retired.